Clean code

+++ R/modeltests.R
@@ -311,33 +311,29 @@ evaluateRuns <- function(model, mydir, gitPath, compScen, email, mattermostToken
         if (currentRunTime > (1.25 * lastRunTime)) {
           errorList <- c(errorList, "Check runtime! Have some scenarios become slower?")
         }
-        if (compScen && !any(grepl("comp_with_.*.pdf", dir()))) {
-          fullPathToThisRun <- normalizePath(".")
-          fullPathToLastRun <- normalizePath(file.path("..",lastRun))
-          #tmp <- read.report(compmif, as.list = FALSE, showSeparatorWarning = FALSE) # fix: can this read.report and write report be skipped? What happens if scenarios in both mifs have the same name?
-          #write.report(x = collapseNames(tmp), file = "tmp.mif", scenario = paste0(cfg$title, "_ref"), model = model)
-          #if (all(file.exists(miffile, "tmp.mif"))) {
-            message("Calling compareScenarios2 with ", fullPathToThisRun, " and ", fullPathToLastRun)
-            if(!test) {
-              #try(system(paste0('Rscript -e "compareScenarios2(c(\"',miffile,'\", \"tmp.mif\"), mifHist = \"historical.mif\",outputFile = \"comp_with_', lastRun,'.pdf\", quiet = TRUE))"')))
-              outFileName <- paste0("comp_with_", lastRun)
-              cs2com <- paste0(
-                "sbatch --qos=standby",
-                " --job-name=", outFileName,
-                " --comment=compareScenarios2",
-                " --output=", fullPathToThisRun, "/", outFileName, ".out",
-                " --error=", fullPathToThisRun, "/", outFileName, ".out",
-                " --mail-type=END --time=200 --mem-per-cpu=8000",
-                " --wrap=\"Rscript scripts/cs2/run_compareScenarios2.R",
-                " outputDirs=", paste(c(fullPathToThisRun, fullPathToLastRun), collapse = ","),
-                " profileName=default",
-                " outFileName=", outFileName,
-                "; ", "mv ",outFileName, ".pdf ", fullPathToThisRun,
-                "\"")
-              cat(cs2com, "\n")
-              withr::with_dir(cfg$remind_folder, {system(cs2com)})
-            }
-          #}
+        fullPathToThisRun <- normalizePath(".")
+        fullPathToLastRun <- normalizePath(file.path("..",lastRun))
+        if (compScen && 
+            all(file.exists(paste0(c(fullPathToThisRun, fullPathToLastRun), "/REMIND_generic_",cfg$title,".mif"))) &&
+            !any(grepl("comp_with_.*.pdf", dir())) &&
+            !test) {
+          message("Calling compareScenarios2 with ", fullPathToThisRun, " and ", fullPathToLastRun)
+          outFileName <- paste0("comp_with_", lastRun)
+          cs2com <- paste0(
+            "sbatch --qos=standby",
+            " --job-name=", outFileName,
+            " --comment=compareScenarios2",
+            " --output=", fullPathToThisRun, "/", outFileName, ".out",
+            " --error=", fullPathToThisRun, "/", outFileName, ".out",
+            " --mail-type=END --time=200 --mem-per-cpu=8000",
+            " --wrap=\"Rscript scripts/cs2/run_compareScenarios2.R",
+            " outputDirs=", paste(c(fullPathToThisRun, fullPathToLastRun), collapse = ","),
+            " profileName=default",
+            " outFileName=", outFileName,
+            "; ", "mv ",outFileName, ".pdf ", fullPathToThisRun,
+            "\"")
+          cat(cs2com, "\n")
+          withr::with_dir(cfg$remind_folder, {system(cs2com)}) # run_compareScenarios2.R works only if called from the main folder 
         }
       }
       setwd("../")
